+++ leda/components/Tooltip/types.ts
@@ -2,6 +2,62 @@ import React from 'react';
 import { COMPONENTS_NAMESPACES } from '../../constants';
 import { PartialGlobalDefaultTheme } from '../../utils/useTheme';
 
+/**
+ * Type of helper that calculates offsets
+ * based on tooltip position
+ */
+export interface GetTooltipOffsets {
+  (data: {
+    /** DOMRect object of target element */
+    elementRect: DOMRect,
+    /** Position of tooltip against target element */
+    position: TooltipPosition,
+  }): {
+    /** Top offset */
+    top?: number,
+    /** Left offset */
+    left?: number,
+  },
+}
+
+/**
+ * Type of helper that calculates position
+ * of tooltip against target element
+ */
+export interface GetTooltipPosition {
+  (data: {
+    /** Width or height of tooltip's arrow  */
+    arrowSize?: number,
+    /** Position of tooltip against target element */
+    position: TooltipPosition,
+    /** DOMRect object of target element */
+    elementRect: DOMRect,
+    /** DOMRect object of tooltip */
+    tooltipRect: DOMRect,
+  }): TooltipPosition,
+}
+
+/**
+ * Properties of TooltipBody
+ */
+export interface TooltipBodyProps {
+  /** Ref of tooltip element */
+  ref?: React.Ref<HTMLDivElement>,
+  /** Handler of transition event */
+  onTransitionEnd: React.TransitionEventHandler,
+  /** Class names */
+  tooltipClassNames?: string,
+  /** Style object */
+  style: TooltipStyle,
+  /** Title is rendered as ReactNode */
+  title: React.ReactNode,
+}
+
+/**
+ * Position of tooltip against target element
+ */
+export type TooltipPosition = 'top' | 'right' | 'bottom' | 'left' | undefined;
+
 export interface TooltipProps {
   /** Размер стрелки тултипа в px */
   arrowSize?: number,
@@ -21,57 +77,45 @@ export interface TooltipProps {
   transitionTimeout?: number,
 }
 
+/** Current ref of tooltip component */
+export interface TooltipRefCurrent {
+  wrapper: HTMLDivElement | null,
+}
+
+/** Tooltip style object */
 export interface TooltipStyle extends React.CSSProperties {
   opacity: 0 | 1,
   top: number,
   left?: number,
   visibility?: 'hidden',
+  width?: number,
+  height?: number,
 }
 
-export type TooltipPosition = 'top' | 'right' | 'bottom' | 'left' | undefined;
-
-export interface TooltipBodyProps {
-  ref?: React.Ref<HTMLDivElement>,
-  onTransitionEnd: React.TransitionEventHandler,
-  tooltipClassNames?: string,
-  style: TooltipStyle,
-  title: React.ReactNode,
-}
-
-export interface TooltipRefCurrent {
-  wrapper: HTMLDivElement | null,
-}
-
-export interface GetTooltipPosition {
-  (data: {
-    arrowSize?: number,
-    position: TooltipPosition,
-    elementRect: DOMRect,
-    tooltipRect: DOMRect,
-  }): TooltipPosition,
-}
-
-export interface GetTooltipOffsets {
-  (data: {
-    elementRect: DOMRect,
-    position: TooltipPosition,
-  }): {
-    top?: number,
-    left?: number,
-  },
-}
-
+/**
+ * Type of custom hook that manages
+ * state of tooltip
+ */
 export interface UseTooltip {
   (data: {
+    /** Width or height of tooltip's arrow  */
     arrowSize?: number,
+    /** Transition timeout  */
     transitionTimeout?: number,
+    /** Defines is prop isOpen was passed to component */
     initialIsOpen?: boolean,
+    /** Defines is prop position was passed to component. Defaults to "top" */
     initialPosition: TooltipPosition,
+    /** Current ref of target element */
     elementRef?: React.RefObject<Element | undefined>,
+    /** Current ref of tooltip */
     tooltipRef?: React.RefObject<Element | undefined>,
   }): {
+    /** Handler of transition event */
     handleTransitionEnd: React.TransitionEventHandler,
+    /** Tooltip position against target element */
     position: TooltipPosition,
+    /** style object */
     style: TooltipStyle,
   },
 }