+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.32.0] - 2020-10-12
 
 ### Fixed
+- Wizard: Added new div element with class .progress-line to set styles for progress bar
 - Modal: Increased the sm width of the component from 420px to 480px
 - DateRange: Added possibility to set array of empty strings as state 
 - MultiSelect: Fixed selectAll text