+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - DropDownSelect: Removed dropdown opening logic from focus handler to click handler 
+- MultiSelect: Fixed checkbox classes 
 - ToolTip: Fixed calculation of position for selecting the opening side
 - DropZone: Prohibited deleting files in disabled mode