+++ CHANGELOG.md
@@ -6,7 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.33.0] - 2020-10-26
 
+## [0.32.0] - 2020-10-15
 
+### Added
+- DatePicker: Added tests for Input, onEnterPress event, test for pressing the "Backspace", "Esc" and "Tab"
 
 ## [0.32.0] - 2020-10-12